pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "poTranslation"
version = "0.0.3a1"
authors = [
    { name="Steven Guo", email="author@example.com" },
]
description = "This is a simple python library/CLI tool to translate .po files using various APIs."
readme = "README.md"
requires-python = ">=3.7"
classifiers = [
    "Development Status :: 3 - Alpha", 
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: BSD License",
    "Operating System :: OS Independent",
    "Topic :: Software Development :: Localization",
    "Topic :: Text Processing :: Linguistic",
    "Topic :: Utilities",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "click",
    "requests",
    "polib",
    "tqdm",
    "python-dotenv",
    "openai",
]


[project.scripts]
# potranslate="poTranslation.cli:translate"
potranslate="poTranslation:translate"
[project.urls]
Homepage = "https://github.com/StevenGuo42/poTranslation"
Repository = "https://github.com/StevenGuo42/poTranslation.git"
"Issue Tracker" = "https://github.com/StevenGuo42/poTranslation/issues"

